Make run_negation as run_time. Fix practice_data/negation. Some changes in getting_started_negation
