--- practice_data/negation/test.tsv
@@ -1,2 +1,0 @@
-The patient is suffering from <e> knee pain </e> but denies nausea .
-The patient is suffering from knee pain but denies <e> nausea </e> .
